src/tools_com/com_Material/com_touch_display.py
"""
Module com_touch_display - Fonctions de présence, fonctionnement et utilisation du tactile de l'écran
"""

import logging

logger = logging.getLogger(__name__)

def check_touch_display_presence():
    """
    Vérifie la présence du tactile de l'écran.

    :return: True si le tactile de l'écran est présent, False sinon.
    :rtype: bool
    """
    try:
        # Vérification de la présence du tactile de l'écran ici
        presence = True  # Exemple de réponse (à adapter)
        if presence:
            logger.info("Touch display is present.")
        else:
            logger.info("Touch display is not present.")
        return presence
    except Exception:
        logger.error("Failed to check touch display presence.")
        return False

def start_touch_display():
    """
    Démarre le tactile de l'écran pour la saisie tactile.

    :return: True si le tactile de l'écran a démarré avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Démarrage du tactile de l'écran ici
        started = True  # Exemple de réponse (à adapter)
        if started:
            logger.info("Touch display started.")
        else:
            logger.error("Failed to start touch display.")
        return started
    except Exception:
        logger.error("Failed to start touch display.")
        return False

def read_touch_input():
    """
    Lit et traite les entrées tactiles de l'écran.

    :return: True si les entrées tactiles ont été lues avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Lecture et traitement des entrées tactiles de l'écran ici
        input_read = True  # Exemple de réponse (à adapter)
        if input_read:
            logger.info("Touch input read and processed.")
        else:
            logger.error("Failed to read touch input.")
        return input_read
    except Exception:
        logger.error("Failed to read touch input.")
        return False
